Add Register folder for UpdateFigureDataEvent
